+++ selfdrive/controls/lib/pid.py
@@ -2,7 +2,7 @@ import numpy as np
 from numbers import Number
 
 class PIDController:
-  def __init__(self, k_p, k_i, k_f=1., k_d=0., pos_limit=1e308, neg_limit=-1e308, rate=100):
+  def __init__(self, k_p, k_i, k_f=0., k_d=0., pos_limit=1e308, neg_limit=-1e308, rate=100):
     self._k_p = k_p
     self._k_i = k_i
     self._k_d = k_d
@@ -16,7 +16,7 @@ class PIDController:
 
     self.set_limits(pos_limit, neg_limit)
 
-    self.i_dt = 1.0 / rate
+    self.i_rate = 1.0 / rate
     self.speed = 0.0
 
     self.reset()
@@ -46,18 +46,21 @@ class PIDController:
 
   def update(self, error, error_rate=0.0, speed=0.0, feedforward=0., freeze_integrator=False):
     self.speed = speed
-    self.p = self.k_p * float(error)
-    self.d = self.k_d * error_rate
-    self.f = self.k_f * feedforward
+    self.p = float(error) * self.k_p
+    self.f = feedforward * self.k_f
+    self.d = error_rate * self.k_d
 
-    if not freeze_integrator:
-      i = self.i + self.k_i * self.i_dt * error
+    i_candidate = self.i if freeze_integrator else self.i + error * self.k_i * self.i_rate
+    u = self.p + i_candidate + self.d + self.f
+    u_sat = np.clip(u, self.neg_limit, self.pos_limit)
 
-      # Don't allow windup if already clipping
-      test_control = self.p + i + self.d + self.f
-      i_upperbound = self.i if test_control > self.pos_limit else self.pos_limit
-      i_lowerbound = self.i if test_control < self.neg_limit else self.neg_limit
-      self.i = np.clip(i, i_lowerbound, i_upperbound)
+    if u == u_sat:
+      self.i = i_candidate
+    else:
+      if u > self.pos_limit and error < 0:
+        self.i = i_candidate
+      elif u < self.neg_limit and error > 0:
+        self.i = i_candidate
 
     control = self.p + self.i + self.d + self.f
     self.control = np.clip(control, self.neg_limit, self.pos_limit)